requirements.txt
# Dash Framework
dash==2.18.2
dash-bootstrap-components==1.6.0

# Database
psycopg2-binary==2.9.10

# Data Processing
pandas==2.2.3
openpyxl==3.1.5

# HTTP Requests
requests==2.32.3
